rename es_ES color provider to es
