+++ faker/providers/color/es_CL/__init__.py
@@ -1,6 +1,7 @@
-from ..es_ES import Provider as ColorProvider
+from ..es import Provider as ColorProvider
 
 localized = True
+
 
 class Provider(ColorProvider):
     pass
